Move verify token endpoint to auth
