+++ backend/src/main/java/com/travelPlanWithAccounting/service/dto/auth/VerifyTokenRequest.java
@@ -1,0 +1,13 @@
+package com.travelPlanWithAccounting.service.dto.auth;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.Data;
+
+@Data
+public class VerifyTokenRequest {
+  @NotNull private VerifyTokenType tokenType; // ACCESS or REFRESH
+  @NotBlank private String token; // JWT 或 RT 明文
+  private String clientId; // REFRESH 可帶（預設 "web"）
+}
+
