+++ backend/src/main/java/com/travelPlanWithAccounting/service/dto/auth/VerifyTokenType.java
@@ -1,0 +1,7 @@
+package com.travelPlanWithAccounting.service.dto.auth;
+
+public enum VerifyTokenType {
+  ACCESS,
+  REFRESH
+}
+
