--- backend/src/main/java/com/travelPlanWithAccounting/service/dto/member/VerifyTokenType.java
@@ -1,3 +1,0 @@
-package com.travelPlanWithAccounting.service.dto.member;
-
-public enum VerifyTokenType { ACCESS, REFRESH }
